app-toolkit: factor WithProperties test helper out of plugin-markdown
Moves the root-collection/SpaceProperties seeding wrapper to a new
@dxos/app-toolkit/testing subpath (CollectionModel lives in app-toolkit)
and updates all importers; removes the core->plugin testing import from
assistant-toolkit.
