+++ packages/plugins/plugin-markdown/src/operations/versioning.test.ts
@@ -6,6 +6,7 @@ import { describe, expect, it } from '@effect/vitest';
 import * as Effect from 'effect/Effect';
 
 import { AssistantTestLayer } from '@dxos/agent-runtime/testing';
+import { WithProperties } from '@dxos/app-toolkit/testing';
 import { SpaceProperties } from '@dxos/client-protocol';
 import { Operation } from '@dxos/compute';
 import { Collection, Database, Feed, Ref, URI } from '@dxos/echo';
@@ -13,8 +14,6 @@ import { TestHelpers } from '@dxos/effect/testing';
 import { invariant } from '@dxos/invariant';
 import { Text } from '@dxos/schema';
 import { HasSubject } from '@dxos/types';
-
-import { WithProperties } from '#testing';
 
 import { Markdown, MarkdownOperation } from '../types';
 import { MarkdownOperationHandlerSet } from './index';
