+++ packages/plugins/plugin-markdown/src/testing.ts
@@ -3,13 +3,11 @@
 //
 
 import * as Toolkit from '@effect/ai/Toolkit';
-import * as Effect from 'effect/Effect';
 
 import { AssistantTestLayer } from '@dxos/agent-runtime/testing';
-import { AppAnnotation } from '@dxos/app-toolkit';
 import { SpaceProperties } from '@dxos/client-protocol';
 import { Skill } from '@dxos/compute';
-import { Annotation, Collection, Database, Feed, Obj, Ref } from '@dxos/echo';
+import { Collection, Feed } from '@dxos/echo';
 import { HasSubject } from '@dxos/types';
 
 import { MarkdownOperationHandlerSet } from '#operations';
@@ -21,27 +19,6 @@ import { Markdown } from '#types';
 // which references React-surface capabilities that are intentionally omitted
 // from `capabilities/node.ts`.
 export * from '#plugin';
-
-// TODO(wittjosiah): Factor out.
-export const WithProperties = <A, E, R>(effect: Effect.Effect<A, E, R>): Effect.Effect<A, E, R | Database.Service> =>
-  Effect.zipRight(
-    Effect.gen(function* () {
-      const collection = Collection.make({ objects: [] });
-      const properties = Obj.make(SpaceProperties, {});
-      yield* Database.add(collection);
-      yield* Database.add(properties);
-      // Both entities are in the DB before setting the annotation so Database.load
-      // works in CollectionModel.add (which uses the Effect DB context, not Ref.load).
-      Obj.update(properties, (properties) => {
-        const meta = Obj.getMeta(properties);
-        if (!meta.annotations) {
-          meta.annotations = {};
-        }
-        Annotation.setDictionary(meta.annotations, AppAnnotation.RootCollectionAnnotation, Ref.make(collection));
-      });
-    }),
-    effect,
-  );
 
 export const testToolkit = Toolkit.empty as Toolkit.Toolkit<any>;
