+++ packages/sdk/app-toolkit/vite.config.ts
@@ -9,6 +9,7 @@ export default defineConfig({
     'index': 'src/index.ts',
     'app-framework/AppActivationEvents': 'src/app-framework/AppActivationEvents.ts',
     'echo/Query': 'src/echo/Query.ts',
+    'testing': 'src/testing/index.ts',
     'ui': 'src/ui/index.ts',
   },
   jsx: 'react',
